requirements.txt
# pip install torch==2.2.0 torchvision==0.17.0 torchaudio==2.2.0 --index-url https://download.pytorch.org/whl/rocm5.6

matplotlib==3.9.1
h5py
librosa
torchlibrosa
huggingface-hub
timm==0.4.5
torch-audiomentations
audiomentations
transformers
wget
wandb
